MS2_storage/requirements.txt
boto3
pandas
psycopg2-binary
minio